+++ lib/decidim/transparent_trash.rb
@@ -6,6 +6,7 @@ require "decidim/transparent_trash/admin_engine"
 require "decidim/transparent_trash/component"
 # Extends
 require "decidim/transparent_trash/extends/unpublish_initiative"
+require "decidim/transparent_trash/extends/initiatives_permissions"
 
 module Decidim
   # This namespace holds the logic of the `TransparentTrash` component. This component
